reporter: move pkg samples from internal to public
https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/245 refactored code and introduced a configuration option to an internal element. Promote package samples from an internal package to public again, so that ExtraSampleAttrProd can be configured again.

Fixes https://github.com/open-telemetry/opentelemetry-ebpf-profiler/issues/280
