+++ reporter/iface.go
@@ -9,7 +9,7 @@ import (
 
 	"go.opentelemetry.io/ebpf-profiler/libpf"
 	"go.opentelemetry.io/ebpf-profiler/process"
-	"go.opentelemetry.io/ebpf-profiler/reporter/internal/samples"
+	"go.opentelemetry.io/ebpf-profiler/reporter/samples"
 )
 
 // Reporter is the top-level interface implemented by a full reporter.
@@ -32,8 +32,6 @@ type Reporter interface {
 	GetMetrics() Metrics
 }
 
-type TraceEventMeta = samples.TraceEventMeta
-
 type TraceReporter interface {
 	// ReportFramesForTrace accepts a trace with the corresponding frames
 	// and caches this information before a periodic reporting to the backend.
@@ -41,12 +39,12 @@ type TraceReporter interface {
 
 	// ReportCountForTrace accepts a hash of a trace with a corresponding count and
 	// caches this information before a periodic reporting to the backend.
-	ReportCountForTrace(traceHash libpf.TraceHash, count uint16, meta *TraceEventMeta)
+	ReportCountForTrace(traceHash libpf.TraceHash, count uint16, meta *samples.TraceEventMeta)
 
 	// ReportTraceEvent accepts a trace event (trace metadata with frames and counts)
 	// and caches it for reporting to the backend. It returns true if the event was
 	// enqueued for reporting, and false if the event was ignored.
-	ReportTraceEvent(trace *libpf.Trace, meta *TraceEventMeta)
+	ReportTraceEvent(trace *libpf.Trace, meta *samples.TraceEventMeta)
 
 	// SupportsReportTraceEvent returns true if the reporter supports reporting trace events
 	// via ReportTraceEvent().
