+++ reporter/samples/attrmgr_test.go
